requirements.txt
Django==4.1.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-webpack-loader==1.6.0
